changed the folder structure
