+++ src/components/Menu/Men.tsx
@@ -1,9 +1,9 @@
 import { Menu } from '@headlessui/react';
 import { FC } from 'react';
 import Link from 'next/link';
-import Feature from '@/components/buttons/Menu/Feature';
+import Feature from '@/components/Menu/Feature';
 
-const Women: FC<{ menu: string }> = ({ menu }) => {
+const Men: FC<{ menu: string }> = ({ menu }) => {
   const Trending = [
     {
       name: 'New Release',
@@ -162,29 +162,53 @@ const Women: FC<{ menu: string }> = ({ menu }) => {
       name: 'Gymshark X Anna',
       href: '/',
     },
+    {
+      name: 'React',
+      href: '/',
+    },
+    {
+      name: 'Arrival',
+      href: '/',
+    },
+    {
+      name: '315 collection',
+      href: '/',
+    },
+    {
+      name: 'Critical',
+      href: '/',
+    },
   ];
 
   const Gym = [
     {
-      name: 'Conditioning Essentials',
+      name: 'Conditioning ',
       href: '/',
     },
     {
-      name: 'Lifting Essentials',
+      name: 'Lifting ',
       href: '/',
     },
     {
       name: 'Reset Day',
       href: '/',
     },
-  ];
-  const WomenSale = [
     {
-      name: 'All Womens Sale',
+      name: 'Running',
       href: '/',
     },
     {
-      name: '50% Off And Above',
+      name: 'Pump Cover',
+      href: '/',
+    },
+  ];
+  const WomenSale = [
+    {
+      name: 'Sweat',
+      href: '/',
+    },
+    {
+      name: 'Sweat',
       href: '/',
     },
   ];
@@ -192,50 +216,11 @@ const Women: FC<{ menu: string }> = ({ menu }) => {
   return (
     <>
       <section
-        className={`absolute left-20 z-20  w-[100%]  rounded-lg bg-white px-6 py-6  `}
+        className={`absolute left-20 z-20 w-[100%]  rounded-lg  bg-white px-6 py-6  `}
       >
         <div className='grid h-[500px] grid-flow-col gap-20'>
           <div>
-            <h5 className='font-bold uppercase'>Trending</h5>
-            {Trending.map((item, i) => {
-              return (
-                <Menu.Item key={i}>
-                  <Link href={item.href}>
-                    <div
-                      className='flex items-baseline justify-between  text-[#666]
- hover:text-[#000]'
-                    >
-                      <h1 className='text-sm	font-semibold	 leading-6  '>
-                        {item.name}
-                      </h1>
-                    </div>
-                  </Link>
-                </Menu.Item>
-              );
-            })}
-            <div className='mt-10'>
-              <h5 className='font-bold uppercase'>Women's sales</h5>
-              {WomenSale.map((item, i) => {
-                return (
-                  <Menu.Item key={i}>
-                    <Link href={item.href}>
-                      <div
-                        className='flex items-baseline	
-  text-[#666] hover:text-[#000]  '
-                      >
-                        <h1 className='text-sm	font-semibold	 leading-6    '>
-                          {item.name}
-                        </h1>
-                      </div>
-                    </Link>
-                  </Menu.Item>
-                );
-              })}
-            </div>
-          </div>
-
-          <div>
-            <h5 className='font-bold uppercase'>ProductS</h5>
+            <h5 className=' font-bold uppercase'>Product</h5>
             {Product.map((item, i) => {
               return (
                 <Menu.Item key={i}>
@@ -253,9 +238,47 @@ const Women: FC<{ menu: string }> = ({ menu }) => {
               );
             })}
           </div>
+          <div>
+            <div>
+              <h5 className='font-bold uppercase'>Men's sales</h5>
+              {WomenSale.map((item, i) => {
+                return (
+                  <Menu.Item key={i}>
+                    <Link href={item.href}>
+                      <div
+                        className='flex items-baseline	
+  text-[#666] hover:text-[#000]  '
+                      >
+                        <h1 className='text-sm	font-semibold	 leading-6    '>
+                          {item.name}
+                        </h1>
+                      </div>
+                    </Link>
+                  </Menu.Item>
+                );
+              })}
+            </div>
+            <h5 className='mt-4 font-bold uppercase'>Trending</h5>
+            {Trending.map((item, i) => {
+              return (
+                <Menu.Item key={i}>
+                  <Link href={item.href}>
+                    <div
+                      className='flex items-baseline justify-between  text-[#666]
+ hover:text-[#000]'
+                    >
+                      <h1 className='text-sm	font-semibold	 leading-6  '>
+                        {item.name}
+                      </h1>
+                    </div>
+                  </Link>
+                </Menu.Item>
+              );
+            })}
+          </div>
 
           <div>
-            <h5 className='font-bold uppercase'>Collections</h5>
+            <h5 className='font-bold uppercase'>Collection</h5>
             {Collection.map((item, i) => {
               return (
                 <Menu.Item key={i}>
@@ -275,7 +298,7 @@ const Women: FC<{ menu: string }> = ({ menu }) => {
           </div>
 
           <div>
-            <h5 className='font-bold uppercase'>Back to Gym Essentials</h5>
+            <h5 className='font-bold uppercase'>Back to Gym Essential</h5>
             {Gym.map((item, i) => {
               return (
                 <Menu.Item key={i}>
@@ -301,4 +324,4 @@ const Women: FC<{ menu: string }> = ({ menu }) => {
     </>
   );
 };
-export default Women;
+export default Men;
